+++ src/game/player/elements/say/Nametag.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { useSentenceContext } from "./context";
+import { useDialogContext } from "./context";
 import { toHex } from "@lib/util/data";
 import { useGame } from "@lib/game/nlcore/common/core";
 
@@ -7,17 +7,17 @@ export default function Nametag({
     ...props
 }: Readonly<React.HTMLAttributes<HTMLDivElement>>) {
     const {
-        sentence,
-    } = useSentenceContext();
+        config,
+    } = useDialogContext();
     const game = useGame();
 
     return (
         <>
             <div {...props}>
                 <span style={{
-                    color: toHex(sentence.config.character?.config.color || game.config.defaultNametagColor),
+                    color: toHex(config.action.character?.config.color || game.config.defaultNametagColor),
                 }}>
-                    {sentence.config.character?.state.name}
+                    {config.action.character?.state.name}
                 </span>
             </div>
         </>