+++ src/game/player/elements/say/context.ts
@@ -1,39 +1,14 @@
-import { GameState } from "@lib/game/nlcore/common/game";
-import { SayElementProps } from "./type";
 import React from "react";
-import { Sentence } from "@core/elements/character/sentence";
-import { Word } from "@core/elements/character/word";
-import { Pausing } from "@core/elements/character/pause";
+import { DialogState } from "./UIDialog";
 
-type SayContext = SayElementProps;
+type DialogContext = DialogState;
 
-export const SayContext = React.createContext<SayContext | null>(null);
+export const DialogContext = React.createContext<DialogContext | null>(null);
 
-export function useSayContext() {
-    const context = React.useContext(SayContext);
+export function useDialogContext() {
+    const context = React.useContext(DialogContext);
     if (!context) {
-        throw new Error("useSayContext must be used within a SayContext");
+        throw new Error("useDialogContext must be used within a DialogContext");
     }
     return context;
 }
-
-export interface SentenceContext {
-    sentence: Sentence;
-    gameState: GameState;
-    finished: boolean;
-    useTypeEffect: boolean;
-    onCompleted?: () => void;
-    count: number;
-    words: Word<Pausing | string>[];
-}
-
-export const SentenceContext = React.createContext<SentenceContext | null>(null);
-
-export function useSentenceContext() {
-    const context = React.useContext(SentenceContext);
-    if (!context) {
-        throw new Error("useSentenceContext must be used within a SentenceContext");
-    }
-    return context;
-}
-