+++ src/index.ts
@@ -1,2 +1,3 @@
 export * from "@core/common/core";
 export * from "@core/common/player";
+import "./styles/index.css";